moved db related functions to postgres pkg
